requirements.txt
playwright>=1.44.0
google-genai>=1.0.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
